feat(auth): split AuthLanding component into its own folder
- Moved AuthLanding component and related files into a dedicated AuthLanding/ directory for better organization and maintainability.
- Cleaned up unused test and config files related to AuthLanding.
